+++ Mean.cpp
@@ -7,7 +7,7 @@
 void Mean::calc() {
     float m = 0.;
     for ( auto itr = cell.begin(); itr != cell.end(); ++itr)
-        m += (*itr)->getValue();
+        m += ((*itr)->getValue().toFloat());
     m /= cell.size();
     value = m;
     ptr->setValue(value);
