+++ SumFormula.cpp
@@ -5,9 +5,9 @@
 #include "SumFormula.h"
 
 void SumFormula::calc() {
-    float x = 0.;
+    float x = 0;
     for (auto itr = cell.begin(); itr != cell.end(); ++itr)
-        x += (*itr)->getValue();
+        x += ((*itr)->getValue().toFloat());
     value = x;
     ptr->setValue(value);
